code refactor of provider logic
