+++ lib/widgets/products_grid.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import './product_item.dart';
+import '../providers/products_provider.dart';
+
+class ProductGrid extends StatelessWidget {
+  const ProductGrid({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final productsData = Provider.of<ProductsProvider>(context);
+    final products = productsData.items;
+
+    return GridView.builder(
+      padding: const EdgeInsets.all(10.0),
+      itemCount: products.length,
+      itemBuilder: (context, index) => ProductItem(
+        products[index].id,
+        products[index].title,
+        products[index].description,
+        products[index].imageUrl,
+        products[index].price,
+      ),
+      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+        crossAxisCount: 2,
+        childAspectRatio: 1,
+        crossAxisSpacing: 15,
+        mainAxisSpacing: 10,
+      ),
+    );
+  }
+}
